test: split parallel failure cases
