+++ projects/04-llm-adapter/tests/compare_runner_parallel/failures/test_parallel_any_runner.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from pathlib import Path
 
 import pytest
